% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{anti_join_.grouped_df}
\alias{anti_join_.grouped_df}
\title{Anti-join a grouped_df.}
\usage{
anti_join_.grouped_df(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{passed}
}
\value{
grouped_df
}
\description{
Anti-joins a grouped_df.
}

