% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant.data.frame}
\alias{constant.data.frame}
\title{Identify Constant Features of a Data Frame}
\usage{
\method{constant}{data.frame}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{grouping columns}
}
\value{
data.frame
}
\description{
Returns columns of a data.frame whose values do not vary within subsets defined by columns named in \dots.
}
\examples{
constant(Theoph, Subject) # Wt Dose
}
\seealso{
Other constant: \code{\link{constant.grouped_df}},
  \code{\link{constant}}
}
