% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant.grouped_df}
\alias{constant.grouped_df}
\title{Identify Constant Features of a Grouped Data Frame}
\usage{
\method{constant}{grouped_df}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{grouping columns}
}
\value{
grouped data.frame
}
\description{
Returns columns of a grouped_df whose values do not vary within subsets defined by groups. 
If any grouping arguments (dots) are supplied, existing groups are over-ridden.
}
\seealso{
Other constant: \code{\link{constant.data.frame}},
  \code{\link{constant}}
}
