% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant}
\alias{constant}
\title{Identify Constant Features of an Object}
\usage{
constant(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Identifies constant features of an object.  Generic, with methods for data.frame and grouped_df.
}
\seealso{
Other constant: 
\code{\link{constant.data.frame}()},
\code{\link{constant.grouped_df}()}
}
\concept{constant}
