% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{detect}
\alias{detect}
\title{Sort column subsets.}
\usage{
detect(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{columns to sort}
}
\value{
grouped_df
}
\description{
Sort column subsets.
}
\seealso{
Other util: 
\code{\link{enumerate}()},
\code{\link{itemize}()},
\code{\link{static}()}
}
\concept{util}
\keyword{internal}
