% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{dupGroups.data.frame}
\alias{dupGroups.data.frame}
\alias{dupGroups}
\title{Count records with with duplicate or duplicated values of grouping variables.}
\usage{
\method{dupGroups}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df

logical
}
\description{
Counts records with with duplicate or duplicated values of grouping variables. If b follows a and and is the same, then b is a duplicate, a is duplicated, and both are shown.
}
\seealso{
Other dupGroups: 
\code{\link{dupGroups_generic}}
}
\concept{dupGroups}
