% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{enumerate}
\alias{enumerate}
\title{Count unique combinations of items in specified columns.}
\usage{
enumerate(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{columns to show}
}
\value{
grouped_df
}
\description{
Counts unique combinations of items in specified columns.
}

