% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{filter_.grouped_df}
\alias{filter_.grouped_df}
\title{Filter a grouped_df.}
\usage{
filter_.grouped_df(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{passed}
}
\value{
grouped_df
}
\description{
Filters a grouped_df.
}

