% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{group_by_all}
\alias{group_by_all}
\title{Group by all columns.}
\usage{
group_by_all(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df
}
\description{
Groups by all columns.
}

