% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{ignore}
\alias{ignore}
\title{Drop columns in x that are present in y.}
\usage{
ignore(x, y, ...)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame}

\item{...}{ingored}
}
\value{
data.frame
}
\description{
Drops columns in x that are present in y.
}
\concept{ignore}
