% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{informative.data.frame}
\alias{informative.data.frame}
\title{Drop columns in x that are entirely NA.}
\usage{
\method{informative}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ingored}
}
\value{
data.frame
}
\description{
Drops columns in x that are entirely NA.
}
\seealso{
Other informative: \code{\link{informative}}
}
