% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{informative}
\alias{informative}
\title{Drop columns in x that are entirely NA.}
\usage{
informative(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ingored}
}
\value{
data.frame
}
\description{
Drops columns in x that are entirely NA.
}

