% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{itemize}
\alias{itemize}
\title{Show unique combinations of items in specified columns}
\usage{
itemize(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{columns to show}
}
\value{
grouped_df
}
\description{
Shows unique combinations of items in specified columns (unquoted).
}
\examples{
itemize(mtcars, cyl, gear, carb)
}
\seealso{
Other util: \code{\link{detect}}, \code{\link{enumerate}},
  \code{\link{static}}
}
