% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{key.grouped_df}
\alias{key.grouped_df}
\title{Fetch the key for a grouped_df as character vector}
\usage{
\method{key}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{columns to show}
}
\value{
character
}
\description{
Fetches the key for a grouped_df as character vector
}
