% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{key}
\alias{key}
\title{Fetch the key.}
\usage{
key(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Fetches the key of an object.
}
\seealso{
Other key: 
\code{\link{key.grouped_df}()}
}
\concept{key}
