% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{misplaced.data.frame}
\alias{misplaced.data.frame}
\title{Index records whose relative positions would change if sorted.}
\usage{
\method{misplaced}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{optional grouping columns (named arguments are ignored)}
}
\value{
logical with length nrow(x)
}
\description{
Indexes records whose relative positions would change if sorted, i.e. records that would not have the same nearest neighbors (before and after).  unsorted() returns the records corresponding to this index.
}
\seealso{
\code{\link{na}} \code{\link{dup}}

Other unsorted: 
\code{\link{misplaced}()},
\code{\link{unsorted.data.frame}()},
\code{\link{unsorted}()}
}
\concept{unsorted}
