% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{misplaced}
\alias{misplaced_generic}
\title{Show misplaced elements.}
\usage{
misplaced(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Shows misplaced elements.
}
\seealso{
Other unsorted: 
\code{\link{misplaced.data.frame}()},
\code{\link{unsorted.data.frame}()},
\code{\link{unsorted_generic}}
}
\concept{unsorted}
\keyword{internal}
