% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{mutate_.grouped_df}
\alias{mutate_.grouped_df}
\title{Mutate a grouped_df.}
\usage{
mutate_.grouped_df(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{passed}
}
\value{
grouped_df
}
\description{
Mutates a grouped_df.
}

