% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{na.data.frame}
\alias{na.data.frame}
\alias{na}
\title{Show records with NA values of grouping variables.}
\usage{
\method{na}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Shows records with NA values of grouping variables.
}
\seealso{
Other na: 
\code{\link{na_generic}}
}
\concept{na}
