% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{na.grouped_df}
\alias{na.grouped_df}
\title{Show records with NA values of grouping variables.}
\usage{
\method{na}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df
}
\description{
Shows records with NA values of grouping variables.
}

