% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{naGroups.data.frame}
\alias{naGroups.data.frame}
\title{Index records with NA values of grouping variables.}
\usage{
\method{naGroups}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{optional grouping columns (named arguments are ignored)}
}
\value{
logical
}
\description{
Indexes records with NA values of grouping variables.
}
\seealso{
Other naGroups: 
\code{\link{naGroups}()}
}
\concept{naGroups}
