% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{naGroups.grouped_df}
\alias{naGroups.grouped_df}
\title{Count records with NA values of grouping variables.}
\usage{
\method{naGroups}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
numeric
}
\description{
Counts records with NA values of grouping variables.
}
\seealso{
Other naGroups: 
\code{\link{naGroups}()}
}
\concept{naGroups}
