% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{naGroups}
\alias{naGroups_generic}
\title{Calculate naGroups.}
\usage{
naGroups(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Calculates naGroups.
}
\seealso{
Other naGroups: 
\code{\link{naGroups.data.frame}()}
}
\concept{naGroups}
\keyword{internal}
