% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{singular}
\alias{singular}
\title{Test whether item has only one unique value.}
\usage{
singular(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{ignored}
}
\value{
logical
}
\description{
Tests whether item has only one unique value.
}

