% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{sort.grouped_df}
\alias{sort.grouped_df}
\title{Arrange by groups.}
\usage{
\method{sort}{grouped_df}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{grouped_df}

\item{decreasing}{logical (ignored)}

\item{...}{further sort criteria}
}
\value{
grouped_df
}
\description{
As of 0.5, dplyr::arrange ignores groups. This function gives the old behavior as a method for generic base::sort. Borrowed from Ax3man at https://github.com/hadley/dplyr/issues/1206.
}
\examples{
library(dplyr)
head(sort(group_by(Theoph, Subject, Time)))
}
