% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{spread_.grouped_df}
\alias{spread_.grouped_df}
\title{Unstack a grouped_df.}
\usage{
spread_.grouped_df(data, key_col, value_col, fill = NA, convert = FALSE,
  drop = TRUE)
}
\arguments{
\item{data}{passed}

\item{key_col}{passed}

\item{value_col}{passed}

\item{fill}{passed}

\item{convert}{passed}

\item{drop}{passed}
}
\value{
grouped_df
}
\description{
Unstacks a grouped_df.
}

