% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{static}
\alias{static}
\title{Find unique records for subset of columns with one unique value.}
\usage{
static(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Finds unique records for subset of columns with one unique value.
}

