% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{status.data.frame}
\alias{status.data.frame}
\title{Report status with respect to grouping variables.}
\usage{
\method{status}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{optional grouping columns (named arguments are ignored)}
}
\value{
returns x invisibly (as originally grouped)
}
\description{
Reports status with respect to grouping variables.
}
\examples{
library(dplyr)
status(Theoph)
status(Theoph, Subject)
status(group_by(Theoph, Subject, Time))
}
\seealso{
\code{\link{na}} 
\code{\link{dup}}
\code{\link{unsorted}}
\code{\link{informative}}
\code{\link{ignore}}
\code{\link{itemize}}
\code{\link{enumerate}}
\code{\link{sort.grouped_df}}

Other status: 
\code{\link{status}()}
}
\concept{status}
