% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{unsorted.data.frame}
\alias{unsorted.data.frame}
\alias{unsorted}
\title{Find records whose relative positions would change if sorted.}
\usage{
\method{unsorted}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Finds records whose relative positions would change if sorted, i.e. records that would not have the same nearest neighbors (before and after).
}
\seealso{
\code{\link{na}} \code{\link{dup}}

Other unsorted: 
\code{\link{unsorted_generic}}
}
\concept{unsorted}
