% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{unsorted.grouped_df}
\alias{unsorted.grouped_df}
\title{Find records whose relative positons would change if sorted.}
\usage{
\method{unsorted}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df
}
\description{
Finds records whose relative positons would change if sorted, i.e. records that would not have the same nearest neighbors (before and after).
}
\seealso{
\code{\link{na}} \code{\link{dup}}

Other unsorted: 
\code{\link{unsorted}()}
}
\concept{unsorted}
