% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{unsorted}
\alias{unsorted_generic}
\title{Show unsorted elements.}
\usage{
unsorted(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Shows unsorted elements.
}
\seealso{
\code{\link{unsorted.data.frame}}

Other unsorted: 
\code{\link{unsorted.data.frame}()}
}
\concept{unsorted}
\keyword{internal}
