% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{weak.data.frame}
\alias{weak.data.frame}
\alias{weak}
\title{Show records with NA, duplicate or duplicated values of grouping variables.}
\usage{
\method{weak}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Shows records with NA, duplicate or duplicated values of grouping variables.
}
\seealso{
Other weak: 
\code{\link{weak_generic}}
}
\concept{weak}
