% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{weak.grouped_df}
\alias{weak.grouped_df}
\title{Show records with NA, duplicate or duplicated values of grouping variables.}
\usage{
\method{weak}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df
}
\description{
Shows records with NA, duplicate or duplicated values of grouping variables.
}
\seealso{
Other weak: \code{\link{weak}}
}
