% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant}
\alias{constant}
\title{Identify Constant Features of an Object}
\usage{
constant(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Identifies constant features of an object.  Generic, with method for data.frame.
}
\seealso{
Other constant: 
\code{\link{constant.data.frame}()}
}
\concept{constant}
\keyword{internal}
