% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{detect}
\alias{detect}
\title{Sort column subsets.}
\usage{
detect(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{columns to sort}
}
\value{
grouped_df
}
\description{
Sort column subsets.
}

