% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{dup.grouped_df}
\alias{dup.grouped_df}
\title{Show records with duplicate or duplicated values of grouping variables.}
\usage{
\method{dup}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df
}
\description{
Shows records with duplicate or duplicated values of grouping variables.
}
\examples{
library(dplyr)
dup(group_by(mtcars, mpg))
}
\seealso{
Other dup: 
\code{\link{dup}()}
}
\concept{dup}
