% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{dup}
\alias{dup_generic}
\title{Show duplicate or duplicated elements.}
\usage{
dup(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Shows duplicate or duplicated elements.
}
\seealso{
\code{\link{dup.data.frame}} \code{\link{na}} \code{\link{weak}}  \code{\link{unsorted}}

Other dup: 
\code{\link{dup.data.frame}()}
}
\concept{dup}
\keyword{internal}
