% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{dupGroups.grouped_df}
\alias{dupGroups.grouped_df}
\title{Count records with with duplicate or duplicated values of grouping variables.}
\usage{
\method{dupGroups}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df
}
\description{
Counts records with with duplicate or duplicated values of grouping variables.
}

