% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{informative}
\alias{informative}
\title{Drop columns in x that are entirely NA.}
\usage{
informative(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Drops columns in x that are entirely NA.
}
\examples{
head(Theoph)
Theoph$Dose <- NA
head(informative(Theoph))
}
\seealso{
\code{\link{informative.data.frame}}

Other informative: 
\code{\link{informative.data.frame}()}
}
\concept{informative}
