% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{na}
\alias{na}
\title{Show na elements.}
\usage{
na(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Shows na elements.
}
\seealso{
\code{\link{na.grouped_df}} \code{\link{dup}} \code{\link{weak}} \code{\link{unsorted}}

Other na: \code{\link{na.grouped_df}}
}
