% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{status.grouped_df}
\alias{status.grouped_df}
\title{Report status with respect to grouping variables.}
\usage{
\method{status}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
returns x invisibly
}
\description{
Reports status with respect to grouping variables.
}

