% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrangle.R
\name{unsorted.grouped_df}
\alias{unsorted.grouped_df}
\title{Find records whose relative positons would change if sorted.}
\usage{
\method{unsorted}{grouped_df}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
grouped_df
}
\description{
Finds records whose relative positons would change if sorted.
}

