% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{weak}
\alias{weak_generic}
\title{Show na, duplicate, or duplicated elements.}
\usage{
weak(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Shows na, duplicate, or duplicated elements.
}
\seealso{
\code{\link{weak.data.frame}}

Other weak: 
\code{\link{weak.data.frame}()}
}
\concept{weak}
\keyword{internal}
