% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{weak}
\alias{weak}
\title{Show na, duplicate, or duplicated elements.}
\usage{
weak(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{other arguments}
}
\description{
Shows na, duplicate, or duplicated elements.
}
\seealso{
\code{\link{weak.grouped_df}}

Other weak: 
\code{\link{weak.grouped_df}()}
}
\concept{weak}
